requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.8
dj-rest-auth==6.0.0
Django==5.0.3
django-allauth==0.63.2
django-cors-headers==4.3.1
django-filter==2.4.0
django-import-export==3.3.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
filelock==3.13.1
gunicorn==21.2.0
idna==3.7
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.0
pillow==10.2.0
platformdirs==4.2.0
pycparser==2.22
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
sqlparse==0.4.4
tablib==3.5.0
urllib3==2.2.1
virtualenv==20.25.1
xlrd==2.0.1
xlwt==1.3.0
